Move highlighted items in the language menu to the top.
When you press Ctrl+L, it's annoying to scroll through a long list of languages.
This patch moves the valid languages (determined by lookahead) to the top of
the menu, separated from the rest with a horizontal bar. In general, this means
that entering the language box you desire is much easier.