First attempt at a principled export framework.
The basic idea is that one subclasses helper.Generic and overrides the functions
in there as needed (hopefully walk doesn't need to be overridden; but everything
else might reasonably need to be).

This clearly isn't perfect (newlines aren't always output where one might
expect), but it's much easier to use than the previous approach and allows
sensible handling of nested language boxes.