Eco is not thread-safe, so doing fancy stuff in another thread causes crashes.
The fancy progress bar spawned another thread which did the loading of
languages. I don't fully know if Python/Eco isn't thread safe in this regard
(Python has lots of thread-related gubbins in its eval module, which looks
scary), but this crashed both my machine and Edd's quite reliably on very large
grammars.

For now, simply disable the progress bar when loading a new file/grammar.

If we want to fix this, it would probably be OK to have the progress bar in
another thread, but the main calculation would have to happen in the main
thread. The main thread could then communicate with the progress bar thread
to tell it to shut itself down.