Spew out install instructions when the py module isn't found.
Since it's impossible to search Google for this stupidly named module, it
baffles everyone who encounters the error "can't import module py". Thus we give
specific instructions for how to install it, and point them to the py module's
website as a backup. This isn't ideal, but I can't think of a better way of
doing this short of removing all usage of py.