lLads of new tests + lingering node fix
When adding nodes to a tree, those nodes are added to
the already saved version. Doing an undo with optimisation would skip
reloading some nodes that version hasn't changed, causing inserted nodes
to "survive" the undo. Everytime we alter nodes, we need to create a meta
version that invalidates the current version and causes the subtree to be
recovered (even though the version hasn't really changed)