Refactored version saving and fixed many bugs
Before versions would be saved after the next keypress that happened after
the version timer ran out. This resulted in some hard to debug behaviour.
Now the version savings occurs simply after the timer ran out (i.e. after
the user stopped typing for a while). To avoid asynchronus saving while
the tree is reparsed, the timer starts only after the parsing step is
completed and aborts if a new parsing routine is initiated.

This patch also adds more tests and fixes several bugs, as well
as adding the option to display the parse trees of different versions to
the gui.