Fixed bug where some changes weren't recovered
The reusing of a subtree during reparsing can cause it do get assigned to
a new parent. That means that the old parent looses track of changes
happening in that subtree. When we recover to a previous version which
reloads the old tree structure the subtree may have changes, but since the
old parent wasn't informed of them the recovery method won't reach the
changed subtree.

This fix marks parents that 'loose' their child nodes as changed on an
assumed basis that the subtree might be changed during it's absence from
the old parent. This ensures that the recovery algorithm will go down
that route in case of changes.