Simple multiline comments through grammar and parser modification
Grammars have a new symbol @, which acts like a wildcard. E.g.

comment ::= # @ #

will parse anything it finds on the tree that is between two #'s.
When parsing terminals the parser checks if the wildcard symbol is valid. Unless the current terminal is
the finishing terminal (#) everything found is pushed on the stack without changing states.
If the second # is found, the parser changes its state to after the wildcard (# @. #) and normal
parsing continues. However the amount of terminals parsed between the #'s is counted and stored as it will be needed later to determine how many elements need to be popped from the stack upon reduction.