Fixed bug where deleting a selection is very slow
When incrementally relexing a node, the algorithm stops when a token is relexed to the same value; only, however, if we passed the initial node where relexing started. There was a bug that when the initial startnode was merged with it's left neighbor, the algorithm wouldn't notice that we passed the startnode and thus relex the entire file. This fix alters the check so that a merged startnode also triggers the `past_startnode` flag.