Fixed error recovery bug in optimisation
To improve the choice of a isolation node, I added an optimisation that removes empty nonterminals from the stack if they interfere with an otherwise valid isolation node. This fix makes sure it only removes empty nonterminals, instead of just checking if a node doesn't have children, which is also true for terminals.