Added top-down reuse
This commit adds Wagner's top-down reuse to Eco. Top-down reuse happens
after parsing has finished and allows the reusing of nodes that bottom-up
reuse can't find.

This is especially useful for empty nonterminals which due to a bug in the
retainability algorithm can't currently be optimistically shifted during
parsing and have to be recreated.