Properly set exists flags and calculate node position
This commit fixes a few inconsistencies with the node.exists-flag which can cause problems with the retain algorithms. This commit also fixes the position calculation used for same_text_pos which was inaccurate. We can't use the position from the history as changing a token at the beginning of the file we'd have to update the positions of all other tokens in the entire tree. Luckily, we only need the position within the `same_text_pos` function. When we reach that function we will have visited all changed nodes up to that point so we can calculate their position on the fly during `pass1`.