Fix incremental AST reusing of nodes
The AST consists of AstNode's and ListNode's of which only AstNode's
are currently being reused. This fixes `is_reusable_astnode` to also
reused ListNode's if they are the same as before. This bug was found
when the Python grammar was fixed to return the correct body in the
`suite_loop` AST rule.