Fix #163: Removing language boxes via selection
Previously, removing all contents of a language box by selecting the whole box
and deleting it, would raise an exception. This was because the cursor sometimes
remained assigned to a deleted node, which also kept us from reparsing the
correct parse tree. The problem has been fixed by doing some checks and
adjusting the cursor node if it is assigned to "dead" nodes.