Update parsetree viewer when switching tabs
Previously, once opened, the parse tree window would only show the program
that was active when the window was opened. Switching to another tab, or
creating a new file, didn't update the parse tree window to show the
respective parse tree. This commit fixes this so the parse tree window
always shows the parse tree of the program currently in focus.