Update auto languagebox options when switching tabs
Previously, if we disabled/enabled auto language boxes in one tab, any
already open tabs would not be informed of that change and use the settings
from when the tab was first opened. This commit fixes this behaviour by
updating a tab with the current settings when it gains focus.