Added SQL grammar adapted from SQLite
While this grammar is still incomplete it currently parses roughly 90%
of the 37,000 SQLite tests. The remaining 10% are either tests that are
expected to fail (as we currently don't have a way of distinguishing
those) or they require us to solve some ambiguities within the grammar.

Since SQLite uses their own grammar syntax and parser generator
(lemonparser), their grammar has ways to solve ambiguities that are not
available in Yacc. For example, the %fallback directive allows the parser
to fallback to a specified token if all other tokens fail to parse (which
SQLite uses to solve identifier ambiguities). Another directive, %wildcard
allows the parser to parse any token within a certain rule. Since Yacc
doesn't have these directives we need to find other ways to work around
this in the future. But for the time being this grammar is decent enough
to work with, especially compared to the old one.