Allow more automatic language box opportunities
Previously, in order to find the terminal from where to start matching
text, we'd find the most right terminal within the subtree of the node
after which a language box can be inserted. If that subtree didn't have a
terminal at its most right position, we'd miss out on some opportunities
to insert a language box, e.g.

`x=SELECT * FROM t`

would not be recognised, because of the way the parse tree is structured
(and `=` is not followed by a whitespace). Now, when no terminal at the
most right position can be found, we traverse the subtree to the left of
this position until we find a terminal, or reach the root of the
subtree.