Fix for longest match and lookaheads
This commit fixes two bugs. The first resulted in the wrong lookahead
being applied to tokens, e.g. when the lexer had revert to a shorted match
because the longer one didn't fully match.

The second bug involved multitokens, where, when finding the longest match,
we compared the list-length of the multitoken against another token,
instead of using the sum of its elements.