Be aware of multinodes when auto-indenting lines
Eco automatically indents newlines if the previous line is indented. For
this it needs to check the leading whitespace of the previous line whenever
a newline is entered. Until now, this process wasn't aware of multinodes
which would cause an error if the previous line is for example a multi-line
comment or string.