Try to continue lexing after error
When an error occurs during lexing, we abort lexing and leave the tokens
in the tree as is. However, this can lead to tokens containing newlines
to not be split up (which is necessary for Eco to function properly).
We can fix this problem, by ensure that even if an error occurs before
a newly inserted token, at the very least that token is relexed, which
ensures that newly inserted newlines are broken up properly.