Fix text-search in Eco to scroll to correct location
Introducing multitext nodes broke the editors text-search in that
it didn't properly scroll to the location of the search result.
The reason was that the search ignored multitext nodes (e.g. comments)
and thus missed some newlines contained within them. Using the method
`next_terminal` (which already knows about multitext nodes) fixes this.