Fix cursor after relexing
After relexing nodes may have been changed in a way that the node the
cursor is attached too is no longer accurate. Previously we would repair
the cursor by checking if the cursor's current node and the cursor
position mismatched (e.g. the cursor position is larger than the text
length of the node). Unfortunately, it is difficult to determine if a
node's text length has been increased or decreased. This occasionally
resulted in weird cursor jumps when typing.

A much easier and straight forward solution to this problem is to simply
store the current cursor position before relexing and then restore it
afterwards (while taking into account any added newlines). In theory
this approach is slightly slower than the previous method, however there
is no noticeable difference when running the tests.