Fix incremental lexing bug
The tree lexer (which works on a parse tree instead of a steam of
tokens) needs to remember which nodes it has seen during lexing so that
the incremental lexer can then merge new tokens with these nodes.

When the lexer sees a node and it is not already in the list of seen
nodes, it is added. Unfortunately, the check for this previously used
equality to determine if the node has been seen before, which leads to
errors when two or more equivalent nodes are processed during relexing.
This commit updates the existence check to use identity instead, which
fixes the problem.