Add menu option to remove language boxes
This commit allows users to remove a selected language box via the
editor's menu, which moves its contents into the outside box.

After removing a language box, also make sure to relex tokens just
before the box as they may need to be merged with the tokens from the
box.