Automatically expand existing language boxes
This commit enables the editor to automatically expand loose language
boxes, by moving code following the language box to the inside.
Currently, a language box always tries to contain the maximum amount of
valid code, although an alternative option would be to only expand a box
if the outside code would otherwise be invalid.