Optimise lexer runtime and memory usage
For regexs like OR, greedy-*, etc. we need backtracking, which requires
storing previous lexing states. Previously, we naively just cloned and
stored the entire lexing state, which included the string we have matched
so far as a list of characters. This list can grow very large and since
every further character we match with greedy-* needs to store the lastest
state, this would lead to us cloning this list multiple times over.
Luckily, we only ever backtrack to a smaller state, which means we only
need to store the index of the last state. Backtracking to this state can
then simply be done by deleting all elements after that index.

Note that the treelexer also stores all nodes we have read so far, so we
can apply the same optimisation there as well.

This optimisation reduced a worst-case scenario by 53x (from 16s to 0.3s).