Allow automatic language boxes to split up tokens
Enables the insertion of automatic language boxes, even if that leads to
an existing node needing to be split up. See the changes in test_eco.py
for an example, where a PHP insertion which contains a partial Java
comment `//';` is split up after the `;` to allow the insertion of a
language box.