Allow duplicate language definitions
Previously, the editor would complain if the user defined the same language
in multiple locations. This is especially likely to occur when users define
language compositions and multiple languages embed another language, e.g.
"Python expression". This required the user to then remove the definition
from one of the compositions, only referencing the name of the embeded
language, which is confusing. This change allows the same definition to
occur multiple times without the editor complaining if they are completely
identical. If, however, one of the definition differs while still using
the same name, the editor will as before notify the user about this.