Improve ranking of candidates
Previously, candidates were ranked by the absolute distance they allow
parsing to continue. Since this distance had a cut-off of 10 tokens,
this gave candidates that occur later in the program an unfair
advantage.

We can improve this by taking the candidate that is most likely to parse
furthest (i.e. the candidate whose ending boundary reaches the furthest)
and check how far it allows parsing to continue (keeping a cut-off of 10
tokens). We then check for each of the remaining candidates if they can
parse as far (or further) than that. If they do they are also equally
valid candidates. If not, they are worse and are removed from the list
of candidates presented to the user.