Fix odd automatic language box behaviour
This fixes an incorrect automatic insertion of a language box when
writing PHP+Python. The reason for this behaviour was that we
automatically remove and insert language boxes at the same time, which
means by the time a language box is automatically inserted, a deletion
may have happened earlier which has changed the program in such a way
that this language box is now incorrect.

We can fix this by separating automatical removal and insertion into two
phases (delete first, then remove). If during deletion of a box a
previous error disappears, so does the candidate attached to that box,
which solves the aforementioned problem.