Fix error when changing language box containing error
This fixes a bug which causes Eco to crash when the user attempts to
change the language of a language box while the content of the box
contains a parse error.

The error manifests in ambig_node_reuse when trying to find a nodes old
parent from a previous version. Since the nodes are copied from the old
language their version differs from the new languages parser which is
initialised to 0. We can fix this by setting the new languages parser
version to the same one as the old language.